--- app/src/main/java/com/example/storeapplication/Signin.kt
@@ -1,40 +1,0 @@
-package com.example.storeapplication
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Button
-import androidx.databinding.DataBindingUtil
-import androidx.navigation.NavController
-import androidx.navigation.findNavController
-import androidx.navigation.fragment.findNavController
-import com.example.storeapplication.databinding.FragmentSigninBinding
-
-class Signin : Fragment() {
-
-    private lateinit var binding: FragmentSigninBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        binding = FragmentSigninBinding.inflate(inflater,container,false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        binding.signUpBtn.setOnClickListener {
-            view.findNavController().navigate(R.id.action_signin_to_signupFragment)
-        }
-
-        binding.signInBtn.setOnClickListener {
-            view.findNavController().navigate(R.id.action_signin_to_homeFragment)
-        }
-
-    }
-}
