+++ app/src/main/res/layout/fragment_signin.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Signin">
+        tools:context=".login.Signin">
 
         <ImageView
             android:id="@+id/e_shopImage"
